Fix event bubbling question
When an event is fired, the browser engine looks for an event handler attached to the element that triggered the event, and then bubbles up the tree giving parent elements a chance to handle the event. Propagation can be canceled during any event handler in the tree.

Event propagation is not "looking for events" -- it's looking for handlers to handle the event that was originally triggered.